requirements.txt
streamlit
psycopg2-binary
pandas
threading
paho.mqtt.client